Remove DuckDB Parquet Reader implementation
