+++ velox/benchmarks/tpch/TpchBenchmark.cpp
@@ -117,7 +117,6 @@ DEFINE_bool(
     false,
     "Include custom statistics along with execution statistics");
 DEFINE_bool(include_results, false, "Include results in the output");
-DEFINE_bool(use_native_parquet_reader, true, "Use Native Parquet Reader");
 DEFINE_int32(num_drivers, 4, "Number of drivers");
 DEFINE_string(data_format, "parquet", "Data format");
 DEFINE_int32(num_splits_per_file, 10, "Number of splits per file");
