--- velox/dwio/parquet/tests/duckdb_reader/ParquetTableScanTest.cpp
@@ -1,234 +1,0 @@
-/*
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include <folly/init/Init.h>
-
-#include "velox/dwio/common/tests/utils/DataFiles.h"
-#include "velox/dwio/parquet/RegisterParquetReader.h"
-#include "velox/dwio/parquet/duckdb_reader/ParquetReader.h"
-#include "velox/exec/tests/utils/HiveConnectorTestBase.h"
-#include "velox/exec/tests/utils/PlanBuilder.h"
-#include "velox/expression/ExprToSubfieldFilter.h"
-
-using namespace facebook::velox;
-using namespace facebook::velox::exec;
-using namespace facebook::velox::exec::test;
-using namespace facebook::velox::parquet;
-
-class ParquetTableScanTest : public HiveConnectorTestBase {
- protected:
-  using OperatorTestBase::assertQuery;
-
-  void SetUp() override {
-    HiveConnectorTestBase::SetUp();
-    unregisterParquetReaderFactory();
-    registerParquetReaderFactory(ParquetReaderType::DUCKDB);
-  }
-
-  void TearDown() override {
-    unregisterParquetReaderFactory();
-    HiveConnectorTestBase::TearDown();
-  }
-
-  void assertSelect(
-      std::vector<std::string>&& outputColumnNames,
-      const std::string& sql) {
-    auto plan = PlanBuilder()
-                    .tableScan(getRowType(std::move(outputColumnNames)))
-                    .planNode();
-
-    assertQuery(plan, splits_, sql);
-  }
-
-  void assertSelectWithFilter(
-      std::vector<std::string>&& outputColumnNames,
-      common::test::SubfieldFilters filters,
-      const std::string& sql) {
-    auto rowType = getRowType(std::move(outputColumnNames));
-
-    auto plan = PlanBuilder()
-                    .tableScan(
-                        rowType,
-                        makeTableHandle(std::move(filters)),
-                        allRegularColumns(rowType))
-                    .planNode();
-
-    assertQuery(plan, splits_, sql);
-  }
-
-  void assertSelectWithAgg(
-      std::vector<std::string>&& outputColumnNames,
-      const std::vector<std::string>& aggregates,
-      const std::vector<std::string>& groupingKeys,
-      const std::string& sql) {
-    auto plan = PlanBuilder()
-                    .tableScan(getRowType(std::move(outputColumnNames)))
-                    .singleAggregation(groupingKeys, aggregates)
-                    .planNode();
-
-    assertQuery(plan, splits_, sql);
-  }
-
-  void assertSelectWithFilterAndAgg(
-      std::vector<std::string>&& outputColumnNames,
-      common::test::SubfieldFilters filters,
-      const std::vector<std::string>& aggregates,
-      const std::vector<std::string>& groupingKeys,
-      const std::string& sql) {
-    auto rowType = getRowType(std::move(outputColumnNames));
-
-    auto plan = PlanBuilder()
-                    .tableScan(
-                        rowType,
-                        makeTableHandle(std::move(filters)),
-                        allRegularColumns(rowType))
-                    .singleAggregation(groupingKeys, aggregates)
-                    .planNode();
-
-    assertQuery(plan, splits_, sql);
-  }
-
-  void
-  loadData(const std::string& filePath, RowTypePtr rowType, RowVectorPtr data) {
-    splits_ = {makeSplit(filePath)};
-    rowType_ = rowType;
-    createDuckDbTable({data});
-  }
-
-  std::string getExampleFilePath(const std::string& fileName) {
-    return facebook::velox::test::getDataFilePath(
-        "velox/dwio/parquet/tests/duckdb_reader", "../examples/" + fileName);
-  }
-
-  std::shared_ptr<connector::hive::HiveConnectorSplit> makeSplit(
-      const std::string& filePath) {
-    return makeHiveConnectorSplits(
-        filePath, 1, dwio::common::FileFormat::PARQUET)[0];
-  }
-
- private:
-  RowTypePtr getRowType(std::vector<std::string>&& outputColumnNames) const {
-    std::vector<TypePtr> types;
-    for (auto colName : outputColumnNames) {
-      types.push_back(rowType_->findChild(colName));
-    }
-
-    return ROW(std::move(outputColumnNames), std::move(types));
-  }
-
-  RowTypePtr rowType_;
-  std::vector<std::shared_ptr<connector::ConnectorSplit>> splits_;
-};
-
-TEST_F(ParquetTableScanTest, basic) {
-  loadData(
-      getExampleFilePath("sample.parquet"),
-      ROW({"a", "b"}, {BIGINT(), DOUBLE()}),
-      makeRowVector(
-          {"a", "b"},
-          {
-              makeFlatVector<int64_t>(20, [](auto row) { return row + 1; }),
-              makeFlatVector<double>(20, [](auto row) { return row + 1; }),
-          }));
-
-  // Plain select
-  assertSelect({"a"}, "SELECT a FROM tmp");
-  assertSelect({"b"}, "SELECT b FROM tmp");
-  assertSelect({"a", "b"}, "SELECT a, b FROM tmp");
-  assertSelect({"b", "a"}, "SELECT b, a FROM tmp");
-
-  // With filters
-  assertSelectWithFilter(
-      {"a"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      "SELECT a FROM tmp WHERE a < 3");
-  assertSelectWithFilter(
-      {"b"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      "SELECT b FROM tmp WHERE a < 3");
-  assertSelectWithFilter(
-      {"a", "b"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      "SELECT a, b FROM tmp WHERE a < 3");
-  assertSelectWithFilter(
-      {"b", "a"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      "SELECT b, a FROM tmp WHERE a < 3");
-  assertSelectWithFilter(
-      {"b"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(0)),
-      "SELECT b FROM tmp WHERE a < 0");
-
-  // TODO: Add filter on b after double filters are supported
-
-  // With aggregations
-  assertSelectWithAgg({"a"}, {"sum(a)"}, {}, "SELECT sum(a) FROM tmp");
-  assertSelectWithAgg({"b"}, {"max(b)"}, {}, "SELECT max(b) FROM tmp");
-  assertSelectWithAgg(
-      {"a", "b"}, {"min(a)", "max(b)"}, {}, "SELECT min(a), max(b) FROM tmp");
-  assertSelectWithAgg(
-      {"b", "a"}, {"max(b)"}, {"a"}, "SELECT max(b), a FROM tmp GROUP BY a");
-  assertSelectWithAgg(
-      {"a", "b"}, {"max(a)"}, {"b"}, "SELECT max(a), b FROM tmp GROUP BY b");
-
-  // With filter and aggregation
-  assertSelectWithFilterAndAgg(
-      {"a"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      {"sum(a)"},
-      {},
-      "SELECT sum(a) FROM tmp WHERE a < 3");
-  assertSelectWithFilterAndAgg(
-      {"b"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      {"sum(b)"},
-      {},
-      "SELECT sum(b) FROM tmp WHERE a < 3");
-  assertSelectWithFilterAndAgg(
-      {"a", "b"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      {"min(a)", "max(b)"},
-      {},
-      "SELECT min(a), max(b) FROM tmp WHERE a < 3");
-  assertSelectWithFilterAndAgg(
-      {"b", "a"},
-      common::test::singleSubfieldFilter("a", exec::lessThan(3)),
-      {"max(b)"},
-      {"a"},
-      "SELECT max(b), a FROM tmp WHERE a < 3 GROUP BY a");
-}
-
-TEST_F(ParquetTableScanTest, countStar) {
-  // sample.parquet holds two columns (a: BIGINT, b: DOUBLE) and
-  // 20 rows
-  auto filePath = getExampleFilePath("sample.parquet");
-  auto split = makeSplit(filePath);
-
-  // Output type does not have any columns
-  auto rowType = ROW({}, {});
-  auto plan = PlanBuilder()
-                  .tableScan(rowType)
-                  .singleAggregation({}, {"count(0)"})
-                  .planNode();
-
-  assertQuery(plan, {split}, "SELECT 20");
-}
-
-int main(int argc, char** argv) {
-  testing::InitGoogleTest(&argc, argv);
-  folly::init(&argc, &argv, false);
-  return RUN_ALL_TESTS();
-}
