+++ velox/dwio/parquet/tests/reader/ParquetReaderBenchmark.cpp
@@ -19,7 +19,6 @@
 #include "velox/dwio/common/Statistics.h"
 #include "velox/dwio/common/tests/utils/DataSetBuilder.h"
 #include "velox/dwio/parquet/RegisterParquetReader.h"
-#include "velox/dwio/parquet/duckdb_reader/ParquetReader.h"
 #include "velox/dwio/parquet/reader/ParquetReader.h"
 #include "velox/dwio/parquet/writer/Writer.h"
 #include "velox/exec/tests/utils/TempDirectoryPath.h"
@@ -116,7 +115,6 @@ class ParquetReaderBenchmark {
   }
 
   std::unique_ptr<RowReader> createReader(
-      const ParquetReaderType& parquetReaderType,
       std::shared_ptr<ScanSpec> scanSpec,
       const RowTypePtr& rowType) {
     dwio::common::ReaderOptions readerOpts{pool_.get()};
@@ -124,18 +122,8 @@ class ParquetReaderBenchmark {
         std::make_shared<LocalReadFile>(fileFolder_->path + "/" + fileName_),
         readerOpts.getMemoryPool());
 
-    std::unique_ptr<Reader> reader;
-    switch (parquetReaderType) {
-      case ParquetReaderType::NATIVE:
-        reader = std::make_unique<ParquetReader>(std::move(input), readerOpts);
-        break;
-      case ParquetReaderType::DUCKDB:
-        reader = std::make_unique<duckdb_reader::ParquetReader>(
-            input->getInputStream(), readerOpts);
-        break;
-      default:
-        VELOX_UNSUPPORTED("Only native or DuckDB Parquet reader is supported");
-    }
+    std::unique_ptr<Reader> reader =
+        std::make_unique<ParquetReader>(std::move(input), readerOpts);
 
     dwio::common::RowReaderOptions rowReaderOpts;
     rowReaderOpts.select(
@@ -148,11 +136,10 @@ class ParquetReaderBenchmark {
   }
 
   int read(
-      const ParquetReaderType& parquetReaderType,
       const RowTypePtr& rowType,
       std::shared_ptr<ScanSpec> scanSpec,
       uint32_t nextSize) {
-    auto rowReader = createReader(parquetReaderType, scanSpec, rowType);
+    auto rowReader = createReader(scanSpec, rowType);
     runtimeStats_ = dwio::common::RuntimeStatistics();
 
     rowReader->resetFilterCaches();
@@ -188,7 +175,6 @@ class ParquetReaderBenchmark {
   }
 
   void readSingleColumn(
-      const ParquetReaderType& parquetReaderType,
       const std::string& columnName,
       const TypePtr& type,
       float startPct,
@@ -219,7 +205,7 @@ class ParquetReaderBenchmark {
 
     // Filter range is generated from a small sample data of 4096 rows. So the
     // upperBound and lowerBound are introduced to estimate the result size.
-    auto resultSize = read(parquetReaderType, rowType, scanSpec, nextSize);
+    auto resultSize = read(rowType, scanSpec, nextSize);
 
     // Add one to expected to avoid 0 in calculating upperBound and lowerBound.
     int expected = kNumBatches * kNumRowsPerBatch *
@@ -265,7 +251,6 @@ void run(
   ParquetReaderBenchmark benchmark(disableDictionary);
   BIGINT()->toString();
   benchmark.readSingleColumn(
-      ParquetReaderType::NATIVE,
       columnName,
       type,
       0,
